+++ eliot/_action.py
@@ -87,6 +87,8 @@ class Action(object):
     @ivar _identification: Fields identifying this action.
 
     @ivar _successFields: Fields to be included in successful finish message.
+
+    @ivar _finished: L{True} if the L{Action} has finished, otherwise L{False}.
     """
     def __init__(self, logger, task_uuid, task_level, action_type,
                  serializers=None):
@@ -121,6 +123,7 @@ class Action(object):
                                 "action_type": action_type,
                                 }
         self._serializers = serializers
+        self._finished = False
 
 
     def _incrementMessageCounter(self):
@@ -151,7 +154,7 @@ class Action(object):
         Message(fields, serializer).write(self._logger, self)
 
 
-    def _finish(self, exception=None):
+    def finish(self, exception=None):
         """
         Log the finish message.
 
@@ -166,6 +169,9 @@ class Action(object):
             which case an C{"exception"} field is added with the given
             L{Exception} type and C{"reason"} with its contents.
         """
+        if self._finished:
+            return
+        self._finished = True
         serializer = None
         if exception is None:
             fields = self._successFields
@@ -255,7 +261,7 @@ class Action(object):
                 exception = result.value
             else:
                 exception = None
-            self._finish(exception)
+            self.finish(exception)
             return result
         deferred.addBoth(done)
 
@@ -284,7 +290,7 @@ class Action(object):
         Pop this action off the execution context, log finish message.
         """
         _context.pop()
-        self._finish(exception)
+        self.finish(exception)