+++ docs/source/news.rst
@@ -12,4 +12,4 @@ Incompatible changes from 0.3.0:
 Features:
 
 * Added support for Python 3.3.
-
+* Actions can now be explicitly finished using a public API: ``Action.finish()``.